Massive changes
To push this update the original practice tables will need to be dropped. So will the players. When I restart the server I will apply these charges and allow the database to be repopulated. To verify functionality.  I renamed the practice tablers and added a line to the yaml file.